split into core and CLI packages
