+++ .github/workflows/node.js.yml
@@ -15,14 +15,15 @@ jobs:
 
     strategy:
       matrix:
-        node-version: [12, 14, 16, 18, 19]
+        node-version: [12, 14, 16, 18, 20]
 
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v4
+    - uses: pnpm/action-setup@v4.0.0
     - name: Use Node.js ${{ matrix.node-version }}
-      uses: actions/setup-node@v2
+      uses: actions/setup-node@v4
       with:
         node-version: ${{ matrix.node-version }}
-        cache: 'npm'
-    - run: npm install
-    - run: npm run test:ci
+        cache: 'pnpm'
+    - run: pnpm install
+    - run: pnpm run test:ci
