+++ .gitignore
@@ -2,4 +2,5 @@ tmp
 out
 .coveralls.yml
 coverage
-node_modules
+node_modules/
+dist/
