+++ packages/bin/cli.js
