+++ packages/bin/lib/cli-data.js
