+++ packages/lib/ansi.js
