+++ test/internals/cell.js
@@ -1,6 +1,4 @@
 import TestRunner from 'test-runner'
-import { strict as a } from 'assert'
-import Cell from '../../lib/cell.js'
 
 const tom = new TestRunner.Tom()
