+++ test/internals/columns.js
@@ -1,6 +1,6 @@
 import TestRunner from 'test-runner'
 import { strict as a } from 'assert'
-import Columns from '../../lib/columns.js'
+import Columns from '../../packages/lib/columns.js'
 
 const tom = new TestRunner.Tom()
